+++ docs/README.md
@@ -67,24 +67,18 @@
 
 
 ## 리팩터링
-- [ ] 피드백 
+- [ ] mvc의 역할(폰 캡처 참고)
 - [ ] 함수(또는 메소드)의 길이가 15라인을 넘어가지 않도록 구현한다. (eslint 속성 추가)
   - 함수(또는 메소드)가 한 가지 일만 잘 하도록 구현한다.
-- [ ] mvc의 역할(폰 캡처 참고)
-- [ ] handleProductAdd를 클래스로 관리? 
+- [ ] 객체화
 - [ ] template 요소 따로 빼주기 
 - [ ] 상수화(메시지, 숫자)
 - [ ] resetInput을 인풋창 별로? 
-- [ ] return true, false 피드백 참고해보기 
 - [ ] alert를 views로 옮기기
-- [ ] 중복된 상품
-- [ ] coins가져오는 연산 리팩터링
-- [ ] let coins같은건 객체화 불가능? (따로 객체 파일 빼서?)
-- [ ] 만약 최초 상품 목록이 있을 때 ? (작년 미션 참고)
-- [ ] 상품 구매 버튼 누르는거 이벤트 위임 제대로 익히기 (클래스 넣었을땐 x)
-- [ ] 0원 투입했을 때 
-- [ ] 반환이 적게 됐을 때, 다시 충전 누르면 제대로 보유금액이 더해지지 않는 문제 fix
-- [ ] 삭제 
+- [ ] 상품 구매 탭도 다시 클릭했을 때 가져올 수 있도록 함 
+- [ ] 피드백 
+- [x] 반환이 적게 됐을 때, 다시 충전 누르면 제대로 보유금액이 더해지지 않는 문제 fix
+- [x] 0원 투입했을 때 alert
 - [x] 수량이 0이 됐을 때 (누르면 -1되는것)
 - [x] css 연결 html파일에 하드코딩 
 ## ❗️ 제출 전 확인하기 
@@ -94,4 +88,8 @@
 - [ ] 디버깅 console.log 없는지 확인
 - [ ] npm run test 
 
-
+# 예상
+- [ ] 삭제 
+- [ ] '사이다' 있을 때, 상품추가하면 기존 '사이다'키에 더해지도록 함 (중복된 상품일 때)
+- [ ] 갯수 0 됐을 때, 목록에서 삭제함 
+- [ ] 만약 최초 상품 목록이 있을 때 ? (작년 미션 참고)
