+++ src/js/controllers/handleProductAdd.js
@@ -1,66 +1,30 @@
 import $ from '../utils/dom.js';
-import { resetProductAddInput } from '../views/resetInput.js';
-import printAddedProduct from '../views/printAddedProduct.js';
 import store from '../utils/store.js';
+import { printAddedProduct, resetProductAddInput } from '../views/productAddView.js';
+import { isValidName, isValidPrice, isValidQuantity } from '../models/productAddModel.js';
 
-const isValidName = nameInput => {
-  if (nameInput === '') {
-    alert('상품명이 공백입니다. 다시 입력하세요.');
-    return false;
-  }
-  return true;
-};
+function HandleProductAdd() {
+  // products는 상태값이 변하니까 this로 관리
+  this.products = store.getLocalStorage('products') ? store.getLocalStorage('products') : [];
 
-const isValidPrice = priceInput => {
-  if (priceInput === '') {
-    alert('가격을 입력하지 않았습니다. 다시 입력하세요.');
-    return false;
-  }
-  if (Number(priceInput) < 100) {
-    alert('상품 가격은 100원 이상만 입력 가능합니다. 다시 입력하세요.');
-    return false;
-  }
-  if (Number(priceInput) % 10 !== 0) {
-    alert('상품 가격은 10원으로 나눠져야합니다. 다시 입력하세요.');
-    return false;
-  }
-  return true;
-};
-
-const isValidQuantity = quantityInput => {
-  if (quantityInput === '') {
-    alert('수량이 입력되지 않았습니다. 다시 입력하세요.');
-    return false;
-  }
-  if (Number(quantityInput) < 1) {
-    alert('수량은 1개 이상 입력되어야합니다. 다시 입력하세요.');
-    return false;
-  }
-  return true;
-};
-
-const handleProductAdd = () => {
-  // 초기값 불러오기 (state)
-  const products = store.getLocalStorage('products') ? store.getLocalStorage('products') : [];
-  if (products.length > 1) {
+  if (this.products.length > 1) {
     printAddedProduct();
   }
 
   $('#product-add-button').addEventListener('click', e => {
     e.preventDefault();
+
     const nameInput = $('#product-name-input').value;
     const priceInput = $('#product-price-input').value;
     const quantityInput = $('#product-quantity-input').value;
 
     if (isValidName(nameInput) && isValidPrice(priceInput) && isValidQuantity(quantityInput)) {
-      const addedProduct = { name: nameInput, price: priceInput, quantity: quantityInput };
-      products.push(addedProduct);
-
-      store.setLocalStorage('products', products);
+      this.products.push({ name: nameInput, price: priceInput, quantity: quantityInput });
+      store.setLocalStorage('products', this.products);
       printAddedProduct();
     }
     resetProductAddInput();
   });
-};
+}
 
-export default handleProductAdd;
+export default HandleProductAdd;
