+++ src/js/index.js
@@ -3,7 +3,7 @@ import renderNavbar from './views/renderNavbar.js';
 import renderProductAdd from './views/renderProductAdd.js';
 import renderVendingMachine from './views/renderVendingMachine.js';
 import renderProductPurchase from './views/renderProductPurchase.js';
-import handleProductAdd from './controllers/handleProductAdd.js';
+import HandleProductAdd from './controllers/HandleProductAdd.js';
 import handleVendingMachine from './controllers/handleVendingMachine.js';
 import handleProductPurchase from './controllers/handleProductPurchase.js';
 
@@ -16,7 +16,7 @@ function VendingMachineApp() {
   $('#app').addEventListener('click', e => {
     if (e.target.id === 'product-add-menu') {
       renderProductAdd();
-      handleProductAdd();
+      new HandleProductAdd();
     }
 
     if (e.target.id === 'vending-machine-manage-menu') {
