+++ src/js/views/resetInput.js
@@ -1,11 +1,5 @@
 import $ from '../utils/dom.js';
 
-export const resetProductAddInput = () => {
-  $('#product-name-input').value = '';
-  $('#product-price-input').value = '';
-  $('#product-quantity-input').value = '';
-};
-
 export const resetChargeInput = () => {
   $('#vending-machine-charge-input').value = '';
 };
